aws: explicitly specify the columns to update during an UPSERT
If we don't explicitly specify the columns to be updated during an UPSERT the
resulting statements will include the unique columns as well, which is not
desired, as they are not supposed to change.

Also, return just the ID of the affected rows, instead of the full payload.

This commit also fixes an issue with the `created_at' and `updated_at' columns,
which were automatically set to the same timestamp.

The `created_at' column is populated once only during initial creation and
should not change throughout the lifecycle of the entity.