clients/gardener: refactor the gardener API clients
Summary of changes:

Move the client into a separate package.

VirtualGardenClient() returns an error in case we don't have the virtual-garden
context.

Also, when creating a new viewerkubeconfig we fetch all pages of shoots first,
instead of just the first page.

The Client can be configured using functional options now.