core/config: add support for parsing configuration from multiple files
Being able to parse configuration from multiple files/sources allows us to split
the configuration into multiple files, and override settings, if needed.

This is also useful in cases where we want to keep some parts of the
configuration (e.g. database settings) encrypted at rest, while the having the
rest of the configuration as-is.