aws/models: remove `rel:has-many` tags from AWS models
The AWS models contain data from multiple AWS accounts, where the `name' and
`account_id' of a resource uniquely identify it.

In order to support such models we need to establish our links/relationships
based on the composite unique key, which is derived from the (name, account_id)
tuple.

Usually this would be solved by adding the `account_id' as a second JOIN
condition for the `rel:has-many' tag, however as of today there are a few open
issues in uptrace/bun [1], which are preventing us from using it.

For example when using multiple JOIN conditions for `rel:has-many', such as the
one below for the models.Region:

Zones []*AvailabilityZone `bun:"rel:has-many,join:name=region_name,join:account_id=account_id"`

This results in errors like the one below when trying to load the relationships
using Relation("Zones").

*errors.errorString: bun: has-many relation=Zones does not have base model=Region with id=["997823283084" "ap-south-1"] (check join conditions)

The open issues related to the error above can be seen in [2], [3] and [4].

Because of these open issues we are not able to use `rel:has-many' tags with
multiple JOIN conditions, and for that reason we are removing these tags from
the existing AWS models.

NOTE: These issues do not affect `rel:has-many' when using a *single* JOIN
condition.

The workaround to these limitations as of now is to use raw queries [5], until
these issues are resolved.

[1]: https://github.com/uptrace/bun
[2]: https://github.com/uptrace/bun/issues/950
[3]: https://github.com/uptrace/bun/issues/992
[4]: https://github.com/uptrace/bun/issues/884
[5]: https://bun.uptrace.dev/guide/query-select.html#raw-queries