aws/tasks: refactor `aws:task:collect-subnets` task
Remove the `aws:task:collect-subnets-region` task, since we no longer need it.

The `aws:task:collect-subnets` task now accepts a payload, which can specify for
which account and region collection to be performed.

In the absence of a payload the task will enqueue collection for subnets from
all known regions and account.