clients/gardener: MCMClient needs a client to a seed, not a shoot
The `MCMClient` creates Machine Controller Manager clients to seed clusters, and
internally uses `createGardenConfig`, which tests whether the name of the seed
actually exists, before requesting `viewerkubeconfig` sub-resource.

Since our managed seeds are actually shoots in the virtual garden cluster, this
is a perfectly valid thing to do. However, technically this is incorrect, since
we should be testing against the list of seeds, and not shoots.

On larger environments this creates a performance issue, where getting all the
shoots at once may result in high memory usage.

This commit will now test whether a seed actually exists and issue a
`viewerkubeconfig` sub-resource for a managed seed, which is still a shoot to
gardener.