gcp/tasks: infer gardener shoot cluster for target pool instances
When the GCP Instance is no longer present in GCP and not part of any Gardener
shoot it is not possible to map the GCE Instance with a shoot in order to
determine whether the shoot is hibernated or not.

This change will infer the Gardener shoot from the VM instance name, thus
allowing us to determine the shoot of an instance, even in the absence of
locally present shoot or Gardener machine.