cmd: remove restriction when submitting tasks
Remove the restriction that the task must be registered when submitting new
tasks to the queue.

We may have extension workers, which reside outside of gardener/inventory for
which we don't have the tasks registered with us.

This change allows `inventory` CLI to submit new tasks for such workers as well.