g/tasks: allow collection of shoots via the cluster-scoped API
Remove payload restriction for having a project namespace specified.

This allows us to pass payloads with empty project namespace, which will then
use the cluster-scoped API to collect the shoots. As a result collecting from
large environments can be processed faster when compared to collection against
each project separately.

The behaviour of the `g:task:collect-shoots` task is as follows.

a) No payload specified

In this case tasks are being enqueued for all known projects, just like we do
now. This method allows for dynamic discovery and collection of shoots, as soon
as new projects are populated in the Inventory database.

b) Project name and project namespace specified

Collection happens only against the project namespace provided as part of the
payload.

c) Project name specified, project namespace is empty

In this collection method the task will use the cluster-scoped API in order to
collect all shoots in a single pass.