g/tasks: use shorter names for the metrics
Also, make sure to set metric values to correct values.

The semantic meaning of the "number of resources collected" metric is as follows.

a) A successful API call to the remote datasource was made to fetch the resources
b) Data has been normalized
c) Normalized records have been persisted locally with Inventory

Only then we consider the "number of resources collected" metric to be accurate,
and it's value represents the number of records fetched *and* persisted with
Inventory.