fix(extension-provider-aws): bump extension-provider-aws version to get fix for breaking spec
Bug:
While strict decoding a providerConfig (defined in aws extension),
we encounter a field name change (capabilitySets -> capabilityFlavours).
Even though we don't use it, the strict decoder fails and reports an error.

PR breaking the current functionality:
https://github.com/gardener/gardener-extension-provider-aws/pull/1480

Fix: bump to provider extension aws, so the expencted
cloudprofile.spec.providerconfig has the same schema as the
ones in the dev landscape.